pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keyboard-mouse-ai-agent"
version = "0.1.0"
description = "A background AI agent service for Windows that controls the user interface via keyboard shortcuts and atomic skills."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["automation", "ai", "gui", "windows", "agent", "pyautogui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyautogui>=0.9.54",
    "keyboard>=0.13.5",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "easyocr>=1.7.0",
    "customtkinter>=5.2.0",
    "darkdetect>=0.8.0",
]

# Windows-specific dependencies
[project.optional-dependencies]
windows = [
    "pywin32>=306; sys_platform == 'win32'",
]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pyinstaller>=6.0.0",
    "keyboard-mouse-ai-agent[windows]",
]

[project.scripts]
km-agent = "km_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "google"
line_length = 100
